__sinit_d_com_inf_game_cpp OK (#140)
* mass mng header

* d_cc_d start

* c_cc_d ok

* tmp commit

* almost, reordering of vtable functions

* progress

* more progress

* cleanup, still non matching

* mark all of d_cc_d as nonmatching to OK

* c_cc_s

* c_cc_s OK

* d_cc_mass_s OK

* cleanup

* gameinfo __sinit, not OK

* a bit of cleanup

* format

* remove asm

+++ include/f_op/f_op_actor_mng.h
@@ -72,7 +72,7 @@ struct DOUBLE_POS {
 };
 
 inline s32 fopAcM_GetRoomNo(const fopAc_ac_c* pActor) {
-    return pActor->mCurrent.mRoomNo;
+    return (s8)pActor->mCurrent.mRoomNo;
 }
 
 inline u32 fopAcM_GetID(const void* pActor) {
